requirements.txt
alembic==1.4.2
amqp==5.1.1
attrs
Babel==2.8.0
billiard==3.6.3.0
blinker==1.4
cookiecutter==1.7.3
celery==5.0.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-log==0.3.2
discord
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Alembic==2.0.1
flask-allows
flask_apscheduler
Flask-BabelPlus==2.2.0
Flask-Caching==1.9.0
Flask-DebugToolbar==0.11.0
flask-debugtoolbar-warnings==0.1.0
Flask-Limiter==1.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.3
Flask-Themes2==0.1.5
flask-whooshee==0.7.0
Flask-WTF==0.14.3
flaskbb-plugin-portal==1.1.3
flask_discord
future==0.18.2
idna==2.10
IPython==8.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.5.3
limits==1.5.1
Mako==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
olefile==0.46
Pillow
pluggy==0.13.1
PyGithub
Pygments==2.6.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.32.3
simplejson==3.17.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
Unidecode==1.1.1
urllib3==1.26.20
vine==5.1.0
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.3
yahoo_fin
-e .
-e ./modules/hub
